custom process snapshot library
This replaces ps as the backend of the process widget.
As only the backend changed there are no major changes for the user, the
only one being that since we now generate the trees for the forest mode
outselves we print them a bit fancier.

We also use some of the code from the process snapshot library to speed
up parsing in the CPU widget.