automatic formatting
I put this off because I'm not 100% happy with the style even a fully
configured clang-format can give but it was really neccessary.  However
I've chosen to try indent instead of clang-format this time. It has some
trouble with some of the more macro heavy code like the vector library
(especially its type syntax), required parsing out all the typedefs, and
even causes unstable formatting (I've noticed it switch between 2 ways
of formatting on alternating runs), but I've put the work into getting
it running and configuring it so I will stick with it for now.