add ability to make the process cursor sticky
The process viewer cursor can now be made to stay at its position on the
screen rather than following the process it's on, allowing jumping to
the top twice and then staying there to view the most intensive
programs.
Previously you had to rely on using the tree mode and selecting systemd
which would always be on the top, and rely on the sorting of its
children providing what you were looking for.